text

+++ libui/src/ui_el/ui_el_remove_texture_by_id.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ui_el_remove_texture_by_id.c                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: sbecker <sbecker@student.42.fr>            +#+  +:+       +#+        */
+/*   By: edraugr- <edraugr-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/23 05:38:20 by sbednar           #+#    #+#             */
-/*   Updated: 2019/07/10 04:46:35 by sbecker          ###   ########.fr       */
+/*   Updated: 2019/07/15 01:17:03 by edraugr-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ void	ui_el_remove_texture_by_id(t_ui_el *el, const char *id)
 	p = NULL;
 	while (l)
 	{
-		if (l->content_size == ft_strhash(id))
+		if ((int)l->content_size == ft_strhash(id))
 		{
 			t = l;
 			if (p)
